requirements.txt
pyTelegramBotAPI==4.19.1
python-dotenv==1.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pillow==10.2.0
gunicorn==21.2.0